[FLINK-27064][test] move ArchUnit rules for production code to the source folder and build a ProductCodeArchitectureBase which can be used by other test outside this module and even in the other repo via dependency.
(cherry picked from commit ff20dbc9179989449780b5cb5c673c43315d9f2b)
